// webapp/app/components/Control/Config/ControlForm/types.ts
/*
 * <<
 * Davinci
 * ==
 * Copyright (C) 2016 - 2017 EDP
 * ==
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * >>
 */

import { IControlRelatedItem, IControlRelatedView } from '../../types'
import { IViewModelProps, IViewVariable } from 'app/containers/View/types'

export interface IFlatRelatedItem extends IControlRelatedItem {
  id: number
  name: string
}

export interface IFlatRelatedView extends IControlRelatedView {
  id: number
  name: string
  models: IViewModelProps[]
  variables: IViewVariable[]
}